refactor(row): chnage component name to TableRow
