+++ src/index.ts
@@ -70,11 +70,11 @@ export * from "./core/Notification";
 export { default as Notification } from "./core/Notification";
 export * from "./core/Radio";
 export { default as RadioButton } from "./core/Radio";
-export * from "./core/Row";
-export { default as Row } from "./core/Row";
 export * from "./core/SegmentedControl";
 export { default as SegmentedControl } from "./core/SegmentedControl";
 export * from "./core/styles";
+export * from "./core/TableRow";
+export { default as TableRow } from "./core/TableRow";
 export * from "./core/Tabs";
 export { default as Tabs } from "./core/Tabs";
 export * from "./core/Tag";
